# backend/app/openapi.yaml
openapi: 3.0.3
info:
  title: Personalized Learning Pathways API
  version: "0.1.0"
  description: >
    MVP API for Personalized Learning Pathways (Goal → Plan → Steps, Q&A, Library).
    Auth via Firebase ID token in Authorization header. Source spec: :contentReference[oaicite:0]{index=0}
servers:
  - url: /api
security:
  - FirebaseBearer: []
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: forbidden
            message:
              type: string
              example: Insufficient permissions
            details:
              type: object
              additionalProperties: true
...
    GoalIntakeQuestion:
      type: object
      required: [id, label, type, order, isActive]
      additionalProperties: false
      properties:
        id:
          type: string
          maxLength: 200
          example: background
        label:
          type: string
          maxLength: 200
          example: What is your current experience?
        type:
          type: string
          enum: [text, multi_select]
          example: text
        options:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 120
          example: [Beginner, Intermediate]
        required:
          type: boolean
          default: false
        order:
          type: integer
          minimum: 0
        isActive:
          type: boolean
          default: true
...
    Category:
      type: object
      required: [id, name, slug, type]
      properties:
        id:
          type: string
          example: cat_editing
        name:
          type: string
          example: Editing
        slug:
          type: string
          example: editing
        type:
          type: string
          enum: [questions, library, mixed]
          example: mixed
    CategoriesListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    CreateCategoryRequest:
      type: object
      required: [name, slug, type]
      properties:
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [questions, library, mixed]
    PatchCategoryRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [questions, library, mixed]
    Goal:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          example: goal_video_editor
        title:
          type: string
          example: Become a video editor
        description:
          type: string
          nullable: true
          example: "..."
        intakeQuestions:
          type: array
          items:
            $ref: "#/components/schemas/GoalIntakeQuestion"
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
    GoalsListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
    CreateGoalRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        intakeQuestions:
          type: array
          items:
            $ref: "#/components/schemas/GoalIntakeQuestion"
    PatchGoalRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        intakeQuestions:
          type: array
          items:
            $ref: "#/components/schemas/GoalIntakeQuestion"
    StepTemplate:
      type: object
      required: [id, title, description, materialUrl, isActive]
      properties:
        id:
          type: string
          example: tmpl_cut_basics
        title:
          type: string
          example: Learn basic cuts
        description:
          type: string
          example: Watch lesson and practice
        materialUrl:
          type: string
          format: uri
          example: https://example.com/lesson
        categoryId:
          type: string
          nullable: true
          example: cat_editing
        tags:
          type: array
          items:
            type: string
          example: [cuts, basics]
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    StepTemplatesListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StepTemplate"
    CreateStepTemplateRequest:
      type: object
      required: [title, description, materialUrl]
      properties:
        title:
          type: string
        description:
          type: string
        materialUrl:
          type: string
          format: uri
        categoryId:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        isActive:
          type: boolean
          default: true
    PatchStepTemplateRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
        materialUrl:
          type: string
          format: uri
        categoryId:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
  responses:
    Unauthorized:
      description: Unauthenticated (missing/invalid token)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden (insufficient role)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /admin/categories:
    get:
      tags: [Admin - Settings]
      summary: List categories
      operationId: adminListCategories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriesListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin - Settings]
      summary: Create category
      operationId: adminCreateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/ValidationError"
  /admin/categories/{id}:
    patch:
      tags: [Admin - Settings]
      summary: Update category
      operationId: adminPatchCategory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchCategoryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Admin - Settings]
      summary: Delete category
      operationId: adminDeleteCategory
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/goals:
    get:
      tags: [Admin - Settings]
      summary: List goals
      operationId: adminListGoals
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin - Settings]
      summary: Create goal
      operationId: adminCreateGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/ValidationError"
  /admin/goals/{id}:
    patch:
      tags: [Admin - Settings]
      summary: Update goal
      operationId: adminPatchGoal
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchGoalRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Admin - Settings]
      summary: Delete goal
      operationId: adminDeleteGoal
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/step-templates:
    get:
      tags: [Admin - Settings]
      summary: List step templates
      operationId: adminListStepTemplates
      parameters:
        - in: query
          name: isActive
          schema: { type: boolean }
        - in: query
          name: categoryId
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepTemplatesListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin - Settings]
      summary: Create step template
      operationId: adminCreateStepTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStepTemplateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/ValidationError"
  /admin/step-templates/{id}:
    patch:
      tags: [Admin - Settings]
      summary: Update step template
      operationId: adminPatchStepTemplate
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchStepTemplateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepTemplate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Admin - Settings]
      summary: Delete step template
      operationId: adminDeleteStepTemplate
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...